i32 registers

+++ mips-emulator/src/emulator.rs
@@ -1,6 +1,6 @@
 use std::{u32, usize};
 
-use crate::state::{StateManager, INSTRUCTION_SIZE, REG_SIZE};
+use crate::state::{StateManager, INSTRUCTION_SIZE };
 
 mod run_i;
 mod run_r;
@@ -8,7 +8,7 @@ mod run_j;
 mod slice_instruction;
 
 
-pub fn run(mut v: Vec<u32>, steps: Option<u32>) -> Vec<u32> {
+pub fn run(mut v: Vec<i32>, steps: Option<u32>) -> Vec<i32> {
    
     let steps_number = match steps {
         Some(v) => v,
@@ -18,8 +18,8 @@ pub fn run(mut v: Vec<u32>, steps: Option<u32>) -> Vec<u32> {
     let (registers, instructions, memory) = StateManager::split(&mut v);
     
     let mut virtual_pc: usize = (registers[32]/4) as usize;
-    let mut lo = registers[34];
-    let mut hi = registers[35];
+    let mut lo = registers[34] as u32;
+    let mut hi = registers[35] as u32;
    
     let mut exit: bool = false;
     let mut steps_counter = 0;
@@ -35,10 +35,10 @@ pub fn run(mut v: Vec<u32>, steps: Option<u32>) -> Vec<u32> {
 
         virtual_pc += 1;
         
-        exit = match slice_instruction::get_op(registers[33]) {
-            0 => run_r::run_r(registers[33], registers, &mut lo, &mut hi),
-            2 | 3 => run_j::run_j(registers[33], registers, &mut virtual_pc),
-            _ => run_i::run_i(registers[33], registers, &mut virtual_pc, memory)
+        exit = match slice_instruction::get_op(registers[33] as u32) {
+            0 => run_r::run_r(registers[33] as u32, registers, &mut lo, &mut hi),
+            2 | 3 => run_j::run_j(registers[33] as u32, registers, &mut virtual_pc),
+            _ => run_i::run_i(registers[33] as u32, registers, &mut virtual_pc, memory)
         };
 
         steps_counter += 1;
@@ -48,7 +48,7 @@ pub fn run(mut v: Vec<u32>, steps: Option<u32>) -> Vec<u32> {
 
     }
 
-    registers[32] = (virtual_pc*4) as u32;
+    registers[32] = (virtual_pc*4) as i32;
 
     return v;
